# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Punjab Blood Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login for hospitals and admins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
        "403":
          description: Hospital not approved yet
  /auth/hospital/register:
    post:
      operationId: registerHospital
      tags: [auth]
      summary: Self-register a hospital or blood bank (pending admin approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HospitalRegistration"
      responses:
        "201":
          description: Registration submitted, pending admin approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HospitalRegistrationResponse"
        "409":
          description: Email already registered

  # --- USER AUTH ---
...
  /user/blood-requests:
    get:
      operationId: listUserBloodRequests
      tags: [user]
      summary: List blood requests submitted by the logged-in user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blood requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodRequest"

    post:
      operationId: createBloodRequest
      tags: [user]
      summary: Submit a targeted blood request to a specific hospital or blood bank
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"
  /user/blood-requests/{id}/cancel:
    patch:
      operationId: cancelBloodRequest
      tags: [user]
      summary: Cancel own pending blood request
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

  # --- HOSPITAL: BLOOD REQUESTS ---
  /hospital/blood-requests:
    get:
      operationId: listHospitalBloodRequests
      tags: [hospital]
      summary: List blood requests received by the logged-in hospital
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: Blood requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloodRequest"
  /hospital/blood-requests/{id}:
    patch:
      operationId: respondToBloodRequest
      tags: [hospital]
      summary: Accept, fulfill, or decline a blood request
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestResponseInput"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

  # --- HOSPITAL DASHBOARD ---
...
  /admin/alerts:
    get:
      operationId: listAdminAlerts
      tags: [admin]
      summary: List admin alerts (e.g. pending hospital registrations)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: unreadOnly
          schema: { type: boolean }
      responses:
        "200":
          description: Admin alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminAlert"
        "401":
          description: Unauthorized
  /admin/alerts/unread-count:
    get:
      operationId: getAdminAlertsUnreadCount
      tags: [admin]
      summary: Count of unread admin alerts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAlertsUnreadCount"
        "401":
          description: Unauthorized
  /admin/alerts/{id}/read:
    patch:
      operationId: markAdminAlertRead
      tags: [admin]
      summary: Mark a single admin alert as read
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Alert marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAlert"
        "404":
          description: Not found
  /admin/alerts/read-all:
    patch:
      operationId: markAllAdminAlertsRead
      tags: [admin]
      summary: Mark all admin alerts as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All alerts marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, role]
      properties:
        token: { type: string }
        role: { type: string }
        hospitalId: { type: ["integer", "null"] }
        name: { type: string }
    AuthMessage:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    HospitalRegistration:
      type: object
      required: [name, registrationNumber, type, ownership, city, district, address, contactPerson, contactNumber, email, password]
      properties:
        name: { type: string }
        registrationNumber: { type: string }
        licenseNumber: { type: string }
        type: { type: string }
        ownership: { type: string }
        city: { type: string }
        district: { type: string }
        state: { type: string }
        address: { type: string }
        pinCode: { type: string }
        contactPerson: { type: string }
        contactNumber: { type: string }
        email: { type: string }
        password: { type: string }
        website: { type: string }
        workingHours: { type: string }
    HospitalRegistrationResponse:
      type: object
      required: [message, hospitalId, approvalStatus]
      properties:
        message: { type: string }
        hospitalId: { type: integer }
        approvalStatus: { type: string }

    # --- USER AUTH ---
...
    BloodRequest:
      type: object
      required: [id, userId, hospitalId, bloodGroup, units, urgency, patientName, contactNumber, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        hospitalId: { type: integer }
        hospitalName: { type: ["string", "null"] }
        bloodGroup: { type: string }
        units: { type: integer }
        urgency: { type: string }
        patientName: { type: string }
        contactNumber: { type: string }
        notes: { type: ["string", "null"] }
        status: { type: string }
        hospitalResponse: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    BloodRequestInput:
      type: object
      required: [hospitalId, bloodGroup, units, urgency, patientName, contactNumber]
      properties:
        hospitalId: { type: integer }
        bloodGroup: { type: string }
        units: { type: integer, minimum: 1 }
        urgency: { type: string }
        patientName: { type: string }
        contactNumber: { type: string }
        notes: { type: string }
    BloodRequestResponseInput:
      type: object
      required: [status]
      properties:
        status: { type: string }
        hospitalResponse: { type: string }

    # --- ADMIN USERS ---
...
    AdminAlert:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        entityType: { type: ["string", "null"] }
        entityId: { type: ["integer", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string }
    AdminAlertsUnreadCount:
      type: object
      required: [count]
      properties:
        count: { type: integer }
...