# lib/api-zod/src/generated/types/bloodRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Punjab Blood Connect API
 * OpenAPI spec version: 0.2.0
 */

export interface BloodRequest {
  id: number;
  userId: number;
  hospitalId: number;
  /** @nullable */
  hospitalName?: string | null;
  bloodGroup: string;
  units: number;
  urgency: string;
  patientName: string;
  contactNumber: string;
  /** @nullable */
  notes?: string | null;
  status: string;
  /** @nullable */
  hospitalResponse?: string | null;
  createdAt: string;
  updatedAt?: string;
}