# lib/api-zod/src/generated/types/hospitalRegistration.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Punjab Blood Connect API
 * OpenAPI spec version: 0.2.0
 */

export interface HospitalRegistration {
  name: string;
  registrationNumber: string;
  licenseNumber?: string;
  type: string;
  ownership: string;
  city: string;
  district: string;
  state?: string;
  address: string;
  pinCode?: string;
  contactPerson: string;
  contactNumber: string;
  email: string;
  password: string;
  website?: string;
  workingHours?: string;
}